Move S3Uploader tests to their own file
